execution/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
google-generativeai==0.8.2
python-dotenv==1.0.1
yt-dlp==2024.5.27
SQLAlchemy==2.0.49
aiosqlite==0.22.1
asyncpg==0.30.0
alembic==1.15.2
PyJWT==2.12.1
opencv-python-headless==4.13.0.92
